[vroom] Change exploration_level from SMALLINT to INTEGER

+++ include/drivers/vroom/vroom_driver.h
@@ -46,7 +46,7 @@ extern "C" {
       Vroom_time_window_t *breaks_tws, size_t total_breaks_tws,
       Vroom_matrix_t *matrix_cells_arr, size_t total_cells,
 
-      int16_t exploration_level,
+      int32_t exploration_level,
       int32_t timeout,
       int32_t loading_time,
